+++ crates/holochain/src/core/workflow/initialize_zomes_workflow.rs
@@ -61,7 +61,7 @@ where
     if result == InitResult::Pass {
         let (flushed_actions, _) = workspace
             .source_chain()
-            .flush(network.target_arcs().await?, network.chc())
+            .flush(network.target_arcs().await?)
             .await?;
 
         send_post_commit(