+++ crates/holochain_conductor_config/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+- **BREAKING CHANGE** Removed `chc` feature. Removed `chc_url` parameter from `generate()`. \#5551
+
 ## 0.7.0-dev.11
 
 ## 0.7.0-dev.10